+++ demo/BAS.hald.R
@@ -12,6 +12,10 @@ hald.coef = coefficients(hald.gprior)
 hald.coef
 plot(hald.coef)
 predict(hald.gprior, top=5, se.fit=TRUE)  
+confint(predict(hald.gprior, Hald, estimator="BMA", se.fit=TRUE, top=5), parm="mean")
+predict(hald.gprior, estimator="MPM", se.fit=TRUE)  
+confint(predict(hald.gprior, Hald, estimator="MPM", se.fit=TRUE), parm="mean")
+
 fitted(hald.gprior, estimator="HPM")
 hald.gprior =  bas.lm(Y~ ., data=Hald, n.models=2^4,
                       prior="g-prior", alpha=13, modelprior=uniform(),