+++ src/hyperg.c
@@ -66,19 +66,18 @@ Copyright 1984, 1987, 1988, 2000 by Stephen L. Moshier
 */
 
 #include "mconf.h"
+#include <Rmath.h>
 
 #ifdef ANSIPROT
 extern double exp ( double );
 extern double log ( double );
-extern double gamma ( double );
-extern double lgam ( double );
 extern double fabs ( double );
 double hyp2f0 ( double, double, double, int, double * );
 static double hy1f1p(double, double, double, double *);
 static double hy1f1a(double, double, double, double *);
 double hyperg (double, double, double);
 #else
-double exp(), log(), gamma(), lgam(), fabs(), hyp2f0();
+double exp(), log(), gammafn(), lgammafn(),fabs(), hyp2f0();
 static double hy1f1p();
 static double hy1f1a();
 double hyperg();
@@ -236,23 +235,23 @@ u = -temp * a;
 
 if( b > 0 )
 	{
-	temp = lgam(b);
+	temp = lgammafn(b);
 	t += temp;
 	u += temp;
 	}
 
 h1 = hyp2f0( a, a-b+1, -1.0/x, 1, &err1 );
 
-temp = exp(u) / gamma(b-a);
+temp = exp(u) / gammafn(b-a);
 h1 *= temp;
 err1 *= temp;
 
 h2 = hyp2f0( b-a, 1.0-a, 1.0/x, 2, &err2 );
 
 if( a < 0 )
-	temp = exp(t) / gamma(a);
+	temp = exp(t) / gammafn(a);
 else
-	temp = exp( t - lgam(a) );
+	temp = exp( t - lgammafn(a) );
 
 h2 *= temp;
 err2 *= temp;
@@ -267,7 +266,7 @@ acanc = fabs(err1) + fabs(err2);
 
 if( b < 0 )
 	{
-	temp = gamma(b);
+	temp = gammafn(b);
 	asum *= temp;
 	acanc *= fabs(temp);
 	}